Unify astype for Series and DataFrame.
